Added function documentation and integer casts to removed signed-unsigned comparison warnings

+++ src/AlignmentFilters.cpp
@@ -186,10 +186,7 @@ namespace AlignmentFilters {
     bases = aln->nucleotides.substr(start_index, num_bases);
   }
 
-
-  /*
-    
-   */ 
+ 
   bool HasLargestEndMatches(AlignedRead* aln, const string& ref_seq, int ref_seq_start, int max_external, int max_internal){
     // Extract sequence, start and end coordinates of read after clipping
     string bases;
@@ -197,16 +194,16 @@ namespace AlignmentFilters {
     GetUnclippedInfo(aln, bases, start, end);
 
     // Check that the prefix match is the longest
-    if (start >= ref_seq_start && start < ref_seq_start + ref_seq.size()){
+    if (start >= ref_seq_start && start < ref_seq_start + static_cast<int>(ref_seq.size())){
       int start_index = start - ref_seq_start;
       int start       = max(0, start_index - max_external);
-      int stop        = min((int)(ref_seq.size()-1), start_index + max_internal);
+      int stop        = min(static_cast<int>((ref_seq.size()-1)), start_index + max_internal);
       vector<int> match_counts;
       ZAlgorithm::GetPrefixMatchCounts(bases, ref_seq, start, stop, match_counts);
 
       int align_index = start_index - start;
       int num_matches = match_counts[align_index];
-      for (unsigned int i = 0; i < match_counts.size(); i++){
+      for (int i = 0; i < static_cast<int>(match_counts.size()); i++){
 	if (i == align_index)
 	  continue;
 	if (match_counts[i] >= num_matches)
@@ -215,23 +212,22 @@ namespace AlignmentFilters {
     }
 
     // Check that the suffix match is the longest
-    if (end >= ref_seq_start && end < ref_seq_start + ref_seq.size()){
+    if (end >= ref_seq_start && end < ref_seq_start + static_cast<int>(ref_seq.size())){
       int end_index = end - ref_seq_start;
       int start     = max(0, end_index - max_internal);
-      int stop      = min((int)(ref_seq.size()-1), end_index + max_external);
+      int stop      = min(static_cast<int>(ref_seq.size()-1), end_index + max_external);
       vector<int> match_counts;
       ZAlgorithm::GetSuffixMatchCounts(bases, ref_seq, start, stop, match_counts);
       
       int align_index = end_index - start;
       int num_matches = match_counts[align_index];
-      for (unsigned int i = 0; i < match_counts.size(); i++){
+      for (int i = 0; i < static_cast<int>(match_counts.size()); i++){
 	if (i == align_index)
 	  continue;
 	if (match_counts[i] >= num_matches)
 	  return false;
       }
-    }    
-   
+    }       
     return true;
   }
 }
